+++ src/App.css
@@ -28,6 +28,56 @@
   color: #fbe161;
 }
 
+.button {
+  display: inline-block;
+  border-radius: 4px;
+  background-color: F0FFFF;
+  border: none;
+  color: #000000;
+  text-align: center;
+  font-size: 28px;
+  padding: 20px;
+  width: 300px;
+  transition: all 0.5s;
+  cursor: pointer;
+  margin: 5px auto;
+}
+
+.button span {
+  cursor: pointer;
+  display: inline-block;
+  position: relative;
+  transition: 0.5s;
+}
+
+.button span:after {
+  content: '»';
+  position: absolute;
+  opacity: 0;
+  top: 0;
+  right: -20px;
+  transition: 0.5s;
+}
+
+.button:hover span {
+  padding-right: 25px;
+}
+
+.button:hover span:after {
+  opacity: 1;
+  right: 0;
+}
+
+.btn1{
+  transform:scale(1)
+}
+
+.btn1:active{
+  transform: scale(0.9);
+  background-color: #0000FF;
+}
+
+
 @keyframes App-logo-spin {
   from {
     transform: rotate(0deg);
